requirements.txt
argon2-cffi
pyotp
qrcode[pil]
kivy
kivymd

